Move CRUD module commands into modulelistcommands package
