+++ src/main/java/seedu/address/logic/parser/DeleteModuleParser.java
@@ -3,7 +3,7 @@ package seedu.address.logic.parser;
 import static seedu.address.commons.core.Messages.MESSAGE_INVALID_COMMAND_FORMAT;
 
 import seedu.address.commons.core.index.Index;
-import seedu.address.logic.commands.DeleteModuleCommand;
+import seedu.address.logic.commands.modulelistcommands.DeleteModuleCommand;
 import seedu.address.logic.parser.exceptions.ParseException;
 
 /**
