+++ src/test/java/seedu/address/logic/commands/DeleteModuleCommandTest.java
@@ -13,6 +13,7 @@ import org.junit.jupiter.api.Test;
 
 // import seedu.address.commons.core.Messages;
 import seedu.address.commons.core.index.Index;
+import seedu.address.logic.commands.modulelistcommands.DeleteModuleCommand;
 import seedu.address.model.Model;
 // import seedu.address.model.ModelManager;
 // import seedu.address.model.UserPrefs;
