+++ src/test/java/seedu/address/logic/parser/DeleteModuleParserTest.java
@@ -7,7 +7,7 @@ import static seedu.address.testutil.TypicalIndexes.INDEX_FIRST_PERSON;
 
 import org.junit.jupiter.api.Test;
 
-import seedu.address.logic.commands.DeleteModuleCommand;
+import seedu.address.logic.commands.modulelistcommands.DeleteModuleCommand;
 
 /**
  * As we are only doing white-box testing, our test cases do not cover path variations
