+++ src/test/java/seedu/address/logic/parser/EditModuleParserTest.java
@@ -20,7 +20,7 @@ import static seedu.address.testutil.TypicalIndexes.INDEX_THIRD_PERSON;
 import org.junit.jupiter.api.Test;
 
 import seedu.address.commons.core.index.Index;
-import seedu.address.logic.commands.EditModuleCommand;
+import seedu.address.logic.commands.modulelistcommands.EditModuleCommand;
 //import seedu.address.model.person.Email;
 //import seedu.address.model.person.Name;
 //import seedu.address.model.tag.Tag;
